+++ projects/ontimize-web-ngx-report/src/lib/interfaces/report-renderer.interface.ts
@@ -1,3 +1,6 @@
 export interface IReportRenderer {
   type: string;
+  renderType?: string,
+  trueValue?: string,
+  falseValue?: string
 }