+++ projects/ontimize-web-ngx-report/src/lib/services/index.ts
@@ -1,3 +1,5 @@
 export * from './o-report.service';
 export * from './ontimize-report.service';
 export * from './o-report-store.service';
+export * from './ontimize-report-data-base-provider.service';
+export * from './ontimize-report-data-provider.service';